Renamed area_market_communicator under redis_connections package to area_market.
