+++ src/gsy_e/models/area/redis_external_market_connection.py
@@ -31,7 +31,7 @@ if TYPE_CHECKING:
     from gsy_e.models.area import Area
     from gsy_e.models.market import Market
     from gsy_e.gsy_e_core.redis_connections.aggregator import AggregatorHandler
-    from gsy_e.gsy_e_core.redis_connections.area_market_communicator import ExternalConnectionCommunicator
+    from gsy_e.gsy_e_core.redis_connections.area_market import ExternalConnectionCommunicator
 
 
 class RedisMarketExternalConnection:
