+++ tests/test_d3a_core/test_redis_area_market_communicator.py
@@ -4,12 +4,12 @@ import pytest
 from redis import StrictRedis
 
 from gsy_e.gsy_e_core.redis_connections.aggregator import AggregatorHandler
-from gsy_e.gsy_e_core.redis_connections.area_market_communicator import ExternalConnectionCommunicator
+from gsy_e.gsy_e_core.redis_connections.area_market import ExternalConnectionCommunicator
 
 
 @pytest.fixture(scope="function", autouse=True)
 def strict_redis():
-    with patch("gsy_e.gsy_e_core.redis_connections.area_market_communicator.StrictRedis",
+    with patch("gsy_e.gsy_e_core.redis_connections.area_market.StrictRedis",
                spec=StrictRedis):
 
         yield
@@ -18,7 +18,7 @@ def strict_redis():
 @pytest.fixture(scope="function", autouse=True)
 def aggregator_handler():
     with patch(
-            "gsy_e.gsy_e_core.redis_connections.area_market_communicator.AggregatorHandler",
+            "gsy_e.gsy_e_core.redis_connections.area_market.AggregatorHandler",
             spec=AggregatorHandler):
         yield
